Update migrate bookmarks

+++ db/migrate/20230404092617_create_bookmarks.rb
@@ -2,7 +2,7 @@ class CreateBookmarks < ActiveRecord::Migration[6.1]
   def change
     create_table :bookmarks do |t|
       t.integer :user_id
-      t.integer :book_id
+      t.bigint :book_id
 
       t.timestamps
     end
